On-disk cache management

cache_directory() reports the path to the on-disk cache for a particular census.

cache_info() summarizes file size and last modification time of files in the cache.

cache() returns an object used to manage the cache.

Usage

cache_directory(id = census_id())

cache_info(id = census_id())

cache(id = census_id())

Arguments

id

character(1) unique identifier associated with a particular census snapshot, as returned by census_id().

Value

cache_directory() returns a character(1) file path to the location of the cache associated with the census with id.

cache_info() returns a tibble with file name, size, and 'mtime' (last-modified time). The mtime is used by cache_disk() to manage the size and age of objects in the cache.

cache() returns an object created by cachem::chach_disk() that can be used to query and delete items in the cache.

Details

The on-disk cache is created and managed by the cache_disk() function in the cachem package. Consult that package for details about working with this object.

File names in the cache are a hash of the function arguments and body of which they are a cache; it is not possible to know transparently which file corresponds to which memoized function.

Examples

cache_directory()
#> [1] "/home/runner/.cache/R/CxGcensus/1598cfd"

cache_info()
#> # A tibble: 13 × 4
#>    file                                      size mtime               path      
#>    <chr>                                    <dbl> <dttm>              <chr>     
#>  1 e24e5139c60be57701400cce0d7489ec.rds   3822607 2023-07-23 01:54:07 /home/run…
#>  2 7a466c52241d4482b73edb2dd91bf29c.rds    740139 2023-07-23 01:53:02 /home/run…
#>  3 fd1b8f254a04e22bffb4cebe18894db3.rds       150 2023-07-23 01:52:56 /home/run…
#>  4 1b1959c977e1.duckdb                  383528960 2023-07-23 01:52:54 /home/run…
#>  5 8bdc07ecb56efaa16923c41298cc8f11.rds  62387006 2023-07-23 01:35:57 /home/run…
#>  6 ccebd96da3e645e0e19e0c798c63bde9.rds    580350 2023-07-23 01:35:14 /home/run…
#>  7 3d6ac4c33357d1131f9698dc2a4d6df3.rds       150 2023-07-23 01:35:13 /home/run…
#>  8 8cc56352bcae0d4b167f5034edc8c7ea.rds     49973 2023-07-23 01:35:13 /home/run…
#>  9 d241a2b48c3d0c82fdc3c44b60d1f364.rds    624139 2023-07-23 01:34:54 /home/run…
#> 10 1049df0a902fb40c35d0b5f34b6f8217.rds     25000 2023-07-23 01:34:19 /home/run…
#> 11 c85095c3479d8e571660137241069fa2.rds       261 2023-07-23 01:34:10 /home/run…
#> 12 d35d065f71053587976b6d3bf08c2592.rds       261 2023-07-23 01:34:05 /home/run…
#> 13 17e17a16ed05.duckdb                   25964544 2023-07-23 01:33:56 /home/run…

cache()
#> <cache_disk> <cachem>
#>   Methods:
#>     get(key, missing = missing_)
#>     set(key, value)
#>     exists(key)
#>     keys()
#>     remove(key)
#>     reset()
#>     prune()
#>     size()
#>     destroy()
#>     is_destroyed(throw = FALSE)
#>     info()