Obtain a reference to a CELLxGENE 'SOMA' collection

census() queries CELLxGENE for a particular census.

census_id() reports a unique identifier for a particular census version.

census_names() queries the census for available 'experiments'. In CELLxGENE, experiments correspond to organisms, e.g., "homo_sapiens" or "mus_musculus".

census_versions() queries the CELLxGENE server for available versions. Versions are denoted by release date.

Usage

census(version = "stable", uri = NULL, tiledbsoma_ctx = NULL)

census_id(version = "stable", uri = NULL, tiledbsoma_ctx = NULL)

census_names(version = "stable", uri = NULL, tiledbsoma_ctx = NULL)

census_versions()

Arguments

version

The version (date) of the census to use. version = "stable" indicates the most recent stable release; version = "latest" is the most recent release. Additional dates are available with census_versions().

uri

The uri corresponding to census version; this is usually discovered automatically.

tiledbsoma_ctx

A 'context' providing mostly low-level flags influencing the performance of tiledbsoma. One illustration of this functionality is in the body of the census() function.

Value

census() returns a tiledbsoma::SOMACollection object. Details of the census are available with census()$get_metadata()

census_id() returns the 7-character git commit sha that uniquely identifies the current release of the census.

census_names() returns a character vector of possible values. Use these values in calls to, e.g., feature_data() or observation_data().

census_versions() returns a tibble with a column of available versions (release dates), and a 'status' column indicating the 'stable' and 'latest' versions.

Details

census() is 'memoised', requiring high-latency internet access only on its first use.

Examples

census()
#> <SOMACollection>
#>   uri: s3://cellxgene-data-public/cell-census/2023-05-15/soma/ 
#>   groups: census_data*, census_info* 
census()$get_metadata()
#> $created_on
#> [1] "2023-05-15T22:34:37+00:00"
#> attr(,"key")
#> [1] "created_on"
#> 
#> $git_commit_sha
#> [1] "1598cfd"
#> attr(,"key")
#> [1] "git_commit_sha"
#> 
#> $soma_encoding_version
#> [1] "1"
#> attr(,"key")
#> [1] "soma_encoding_version"
#> 
#> $soma_object_type
#> [1] "SOMACollection"
#> attr(,"key")
#> [1] "soma_object_type"
#> 

census_id()
#> [1] "1598cfd"

census_names()
#> [1] "homo_sapiens" "mus_musculus"

census_versions()
#> # A tibble: 6 × 2
#>   version    status
#>   <date>     <chr> 
#> 1 2023-05-15 stable
#> 2 2023-06-20 NA    
#> 3 2023-06-28 NA    
#> 4 2023-07-03 NA    
#> 5 2023-07-10 NA    
#> 6 2023-07-17 latest

tiledbsoma::show_package_versions()
#> tiledbsoma:    0.99.1
#> tiledb-r:      0.20.1
#> tiledb core:   2.16.0
#> libtiledbsoma: libtiledb=2.16.0
#> R:             R version 4.3.1 (2023-06-16)
#> OS:            Ubuntu 22.04.2 LTS