Discover census datasets, cells, features, and observations

datasets() queries CELLxGENE for datasets used in constructing the census.

summary_cell_counts() reports the facets (e.g., sex) and levels (e.g., male, female) in the census, and the number of cells associated with each facet and level.

feature_data() reports information about features (genes) present in the census.

observation_data() reports information about all cells in the census.

assay_data() queries the census for 'raw' counts for the RNA-seq data corresponding to selected features and columns in a census.

single_cell_experiment() queries the census for assay data corresponding to features and observations, and assembles the result into a SingleCellExperiment. The count data are accessible using SingleCellExperiment::counts().

Usage

datasets(version = "stable", uri = NULL, tiledbsoma_ctx = NULL)

summary_cell_counts(version = "stable", uri = NULL, tiledbsoma_ctx = NULL)

feature_data(organism, version = "stable", uri = NULL, tiledbsoma_ctx = NULL)

observation_data(
  organism,
  version = "stable",
  uri = NULL,
  tiledbsoma_ctx = NULL
)

assay_data(
  organism,
  features,
  observations,
  version = "stable",
  uri = NULL,
  tiledbsoma_ctx = NULL
)

single_cell_experiment(
  organism,
  features,
  observations,
  version = "stable",
  uri = NULL,
  tiledbsoma_ctx = NULL
)

Arguments

version

The version (date) of the census to use. version = "stable" indicates the most recent stable release; version = "latest" is the most recent release. Additional dates are available with census_versions().

uri

The uri corresponding to census version; this is usually discovered automatically.

tiledbsoma_ctx

A 'context' providing mostly low-level flags influencing the performance of tiledbsoma. One illustration of this functionality is in the body of the census() function.

organism

one of the values returned by census_names(), specifically 'homo_sapiens' or 'mus_musculus' at the time of writing this documentation.

features

a tibble, typically derived from feature_data() via filter(), select(), etc., and containing the soma_joinid column.

observations

a tibble, typically derived from observation_data() via filter(), select(), etc., and containing the soma_joinid column.

Value

datasets() returns a tibble with information about the collections and datasets represented in the census.

summary_cell_counts() returns a tibble summarizing the organism, facets (category, e.g., 'sex') and levels (label, e.g., 'female') represented in the data, and unique and total cell counts in each facet and level.

feature_data() returns a tibble with columns describing each feature (gene) in organism.

observation_data() returns a dbplyr-based tibble of cell annotations. An aesthetic problem is that the 'connection' to the database is not available to the user, and duckdb warns that Database is garbage-collected...; this message can be ignored.

assay_data() returns a sparse matrix (dgCMatrix) summarizing counts found in the census for the soma_joinid

columns of the features and observations tibble arguments. The counts are from the 'raw' layer' of 'X' collection of the 'RNA' measurement in the experiment.

Details

datasets(), summary_cell_counts(), an feature_data() are 'memoised' so that they are only expensive on their first use. The 'tibble' returned by these functions is memoised to disk, so that re-using the function is fast even across sessions. See ?cache_info() for details on cache management.

observation_data() is memoised to disk. The data is large (e.g., more than 50 million rows for Homo sapiens) so the initial download can be time-consuming (10's of minutes). During download in an interactive session, the number of 'chunks' and records are displayed; for the 2023-05-15 census of homo_sapiens, there were more than 52 million records (cells) downloaded in 124 chunks.

The data are stored in a 'duckdb' database. The return value can be used via dbplyr for very fast and memory efficient filtering, selection, and summary.

Currently, for assay_data() and single_cell_experiment(), the user must ensure that the features, observations, and assay data are for the same census and organism. Duplicate rows in features and observations are allowed.

single_cell_experiment() requires that the SingleCellExperiment Bioconductor package is installed, e.g., via BiocManager::install("SingleCellExperiment").

Examples

datasets() |>
    glimpse()
#> Rows: 562
#> Columns: 8
#> $ soma_joinid              <int> 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13,…
#> $ collection_id            <chr> "6b701826-37bb-4356-9792-ff41fc4c3161", "4195…
#> $ collection_name          <chr> "Abdominal White Adipose Tissue", "A spatiall…
#> $ collection_doi           <chr> NA, NA, NA, NA, NA, "10.1016/j.cell.2021.12.0…
#> $ dataset_id               <chr> "9d8e5dca-03a3-457d-b7fb-844c75735c83", "a638…
#> $ dataset_title            <chr> "22 integrated samples", "scRNA-seq data - my…
#> $ dataset_h5ad_path        <chr> "9d8e5dca-03a3-457d-b7fb-844c75735c83.h5ad", …
#> $ dataset_total_cell_count <int> 72335, 30789, 714331, 76567, 12510, 167598, 4…

summary_cell_counts() |>
    count(category)
#> # A tibble: 9 × 2
#>   category                    n
#>   <chr>                   <int>
#> 1 all                         2
#> 2 assay                      29
#> 3 cell_type                 830
#> 4 disease                    73
#> 5 self_reported_ethnicity    27
#> 6 sex                         6
#> 7 suspension_type             2
#> 8 tissue                    278
#> 9 tissue_general             88

## number of cells from female, male, and 'unknown' samples in
## humans and mice
summary_cell_counts() |>
    filter(category == "sex") |>
    select(
        organism, label,
        unique_cell_count, total_cell_count
    )
#> # A tibble: 6 × 4
#>   organism     label   unique_cell_count total_cell_count
#>   <chr>        <chr>               <int>            <int>
#> 1 Homo sapiens female           14516846         22513226
#> 2 Homo sapiens male             17097019         28197731
#> 3 Homo sapiens unknown           2145022          3083771
#> 4 Mus musculus female            1066585          1431227
#> 5 Mus musculus male              1655113          2462185
#> 6 Mus musculus unknown            192620           192620

feature_data("mus_musculus")
#> # A tibble: 52,392 × 4
#>    soma_joinid feature_id         feature_name  feature_length
#>          <int> <chr>              <chr>                  <int>
#>  1           0 ENSMUSG00000109644 0610005C13Rik           3583
#>  2           1 ENSMUSG00000108652 0610006L08Rik           2128
#>  3           2 ENSMUSG00000007777 0610009B22Rik            998
#>  4           3 ENSMUSG00000086714 0610009E02Rik           1803
#>  5           4 ENSMUSG00000043644 0610009L18Rik            619
#>  6           5 ENSMUSG00000042208 0610010F05Rik           5226
#>  7           6 ENSMUSG00000020831 0610010K14Rik           1896
#>  8           7 ENSMUSG00000089755 0610012D04Rik            531
#>  9           8 ENSMUSG00000107002 0610012G03Rik           1445
#> 10           9 ENSMUSG00000046683 0610025J13Rik           1121
#> # ℹ 52,382 more rows

mus <- observation_data("mus_musculus")
mus |>
    count(assay, sort = TRUE)
#> # Source:     SQL [9 x 2]
#> # Database:   DuckDB 0.8.1 [unknown@Linux 5.15.0-1041-azure:R 4.3.1//home/runner/.cache/R/CxGcensus/18292c41a745.duckdb]
#> # Ordered by: desc(n)
#>   assay                                n
#>   <chr>                          <int64>
#> 1 10x 3' v3                      1925688
#> 2 10x 3' v2                      1622998
#> 3 Smart-seq2                      221648
#> 4 10x 5' v1                       161525
#> 5 Drop-seq                         80990
#> 6 Smart-seq                        48972
#> 7 10x 3' v1                        21421
#> 8 DroNc-seq                         2768
#> 9 10x 3' transcription profiling      22
mus |>
    filter(grepl("diabetes", disease)) |>
    count(disease, sex, tissue)
#> # Source:   SQL [2 x 4]
#> # Database: DuckDB 0.8.1 [unknown@Linux 5.15.0-1041-azure:R 4.3.1//home/runner/.cache/R/CxGcensus/18292c41a745.duckdb]
#>   disease                  sex    tissue                    n
#>   <chr>                    <chr>  <chr>               <int64>
#> 1 type 2 diabetes mellitus male   islet of Langerhans   99747
#> 2 type 1 diabetes mellitus female islet of Langerhans   39932

features <-
   feature_data("mus_musculus") |>
   ## rows 4, 3, 4 of the tibble
   slice(c(4:3, 4))

observations <-
   observation_data("mus_musculus") |>
   ## first two rows of the tibble
   head(2) |>
   collect()

assay_data(features, observations, organism = "mus_musculus")
#> creating axis and experiment queries
#> retrieving assay_data measurement 'RNA' collection 'X' layer 'raw' as a
#> sparse matrix with 2 x 2 distinct features x observations
#> 3 x 2 sparse Matrix of class "dgCMatrix"
#>   0 1
#> 3 1 .
#> 2 . .
#> 3 1 .

single_cell_experiment("mus_musculus", features, observations)
#> creating axis and experiment queries
#> retrieving assay_data measurement 'RNA' collection 'X' layer 'raw' as a
#> sparse matrix with 2 x 2 distinct features x observations
#> class: SingleCellExperiment 
#> dim: 3 2 
#> metadata(1): census_metadata
#> assays(1): counts
#> rownames: NULL
#> rowData names(4): soma_joinid feature_id feature_name feature_length
#> colnames: NULL
#> colData names(21): soma_joinid dataset_id ... tissue_general
#>   tissue_general_ontology_term_id
#> reducedDimNames(0):
#> mainExpName: NULL
#> altExpNames(0):