Construct a SingleCellExperiment from census data

single_cell_experiment() queries the census for assay data corresponding to features and observations, and assembles the result into a SingleCellExperiment. The count data are accessible using SingleCellExperiment::counts().

Usage

single_cell_experiment(
  organism,
  features,
  observations,
  version = "stable",
  uri = NULL,
  tiledbsoma_ctx = NULL
)

Arguments

organism

one of the values returned by census_names(), specifically 'homo_sapiens' or 'mus_musculus' at the time of writing this documentation.

features

a tibble, typically derived from feature_data() via filter(), select(), etc., and containing the soma_joinid column.

observations

a tibble, typically derived from observation_data() via filter(), select(), etc., and containing the soma_joinid column.

version

The version (date) of the census to use. version = "stable" indicates the most recent stable release; version = "latest" is the most recent release. Additional dates are available with census_versions().

uri

The uri corresponding to census version; this is usually discovered automatically.

tiledbsoma_ctx

A 'context' providing mostly low-level flags influencing the performance of tiledbsoma. One illustration of this functionality is in the body of the census() function.

Details

single_cell_experiment() requires that the SingleCellExperiment Bioconductor package is installed, e.g., via BiocManager::install("SingleCellExperiment").

Examples

features <-
   feature_data("mus_musculus") |>
   ## rows 4, 3, 4 of the tibble
   slice(c(4:3, 4))

observations <-
   observation_data("mus_musculus") |>
   ## first two rows of the tibble
   head(2) |>
   collect()

single_cell_experiment("mus_musculus", features, observations)
#> creating axis and experiment queries
#> retrieving assay_data measurement 'RNA' collection 'X' layer 'raw' as a
#> sparse matrix with 2 x 2 distinct features x observations
#> class: SingleCellExperiment 
#> dim: 3 2 
#> metadata(1): census_metadata
#> assays(1): counts
#> rownames: NULL
#> rowData names(4): soma_joinid feature_id feature_name feature_length
#> colnames: NULL
#> colData names(21): soma_joinid dataset_id ... tissue_general
#>   tissue_general_ontology_term_id
#> reducedDimNames(0):
#> mainExpName: NULL
#> altExpNames(0):