On-disk cache management

cache_directory() reports the path to the on-disk cache for a particular census.

cache_info() summarizes file size and last modification time of files in the cache.

cache() returns an object used to manage the cache.

Usage

cache_directory(id = census_id())

cache_info(id = census_id())

cache(id = census_id())

Arguments

id

character(1) unique identifier associated with a particular census snapshot, as returned by census_id().

Value

cache_directory() returns a character(1) file path to the location of the cache associated with the census with id.

cache_info() returns a tibble with file name, size, and 'mtime' (last-modified time). The mtime is used by cache_disk() to manage the size and age of objects in the cache.

cache() returns an object created by cachem::chach_disk() that can be used to query and delete items in the cache.

Details

The on-disk cache is created and managed by the cache_disk() function in the cachem package. Consult that package for details about working with this object.

File names in the cache are a hash of the function arguments and body of which they are a cache; it is not possible to know transparently which file corresponds to which memoized function.

Examples

cache_directory()
#> [1] "/home/runner/.cache/R/CxGcensus/1598cfd"

cache_info()
#> # A tibble: 0 × 4
#> # ℹ 4 variables: file <chr>, size <dbl>, mtime <dttm>, path <chr>

cache()
#> <cache_disk> <cachem>
#>   Methods:
#>     get(key, missing = missing_)
#>     set(key, value)
#>     exists(key)
#>     keys()
#>     remove(key)
#>     reset()
#>     prune()
#>     size()
#>     destroy()
#>     is_destroyed(throw = FALSE)
#>     info()