Discover census datasets and cell count summaries

datasets() queries CELLxGENE for datasets used in constructing the census.

summary_cell_counts() reports the facets (e.g., sex) and levels (e.g., male, female) in the census, and the number of cells associated with each facet and level.

Usage

datasets(version = "stable", uri = NULL, tiledbsoma_ctx = NULL)

summary_cell_counts(version = "stable", uri = NULL, tiledbsoma_ctx = NULL)

Arguments

version

The version (date) of the census to use. version = "stable" indicates the most recent stable release; version = "latest" is the most recent release. Additional dates are available with census_versions().

uri

The uri corresponding to census version; this is usually discovered automatically.

tiledbsoma_ctx

A 'context' providing mostly low-level flags influencing the performance of tiledbsoma. One illustration of this functionality is in the body of the census() function.

Value

datasets() returns a tibble with information about the collections and datasets represented in the census.

summary_cell_counts() returns a tibble summarizing the organism, facets (category, e.g., 'sex') and levels (label, e.g., 'female') represented in the data, and unique and total cell counts in each facet and level.

Details

datasets(), summary_cell_counts(), an feature_data() are 'memoised' so that they are only expensive on their first use. The 'tibble' returned by these functions is memoised to disk, so that re-using the function is fast even across sessions. See ?cache_info() for details on cache management.

Examples

datasets() |>
    glimpse()
#> Rows: 593
#> Columns: 8
#> $ soma_joinid              <int> 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13,…
#> $ collection_id            <chr> "e2c257e7-6f79-487c-b81c-39451cd4ab3c", "e2c2…
#> $ collection_name          <chr> "Spatial multiomics map of trophoblast develo…
#> $ collection_doi           <chr> "10.1038/s41586-023-05869-0", "10.1038/s41586…
#> $ dataset_id               <chr> "f171db61-e57e-4535-a06a-35d8b6ef8f2b", "ecf2…
#> $ dataset_title            <chr> "donor_p13_trophoblasts", "All donors trophob…
#> $ dataset_h5ad_path        <chr> "f171db61-e57e-4535-a06a-35d8b6ef8f2b.h5ad", …
#> $ dataset_total_cell_count <int> 31497, 67070, 286326, 270855, 167283, 13497, …

summary_cell_counts() |>
    count(category)
#> # A tibble: 9 × 2
#>   category                    n
#>   <chr>                   <int>
#> 1 all                         2
#> 2 assay                      28
#> 3 cell_type                 861
#> 4 disease                    69
#> 5 self_reported_ethnicity    27
#> 6 sex                         6
#> 7 suspension_type             2
#> 8 tissue                    286
#> 9 tissue_general             81

## number of cells from female, male, and 'unknown' samples in
## humans and mice
summary_cell_counts() |>
    filter(category == "sex") |>
    select(
        organism, label,
        unique_cell_count, total_cell_count
    )
#> # A tibble: 6 × 4
#>   organism     label   unique_cell_count total_cell_count
#>   <chr>        <chr>               <int>            <int>
#> 1 Homo sapiens female           13955052         22198837
#> 2 Homo sapiens male             17152628         30950702
#> 3 Homo sapiens unknown           2256562          3251334
#> 4 Mus musculus female            1467473          1832115
#> 5 Mus musculus male              2423438          3230510
#> 6 Mus musculus unknown            192620           192620